CompositorClient: Restructure sources and add a test app
