+++ Source/compositorclient/CMakeLists.txt
@@ -23,67 +23,7 @@ project(compositorclient)
 
 project_version(1.0.0)
 
-set(TARGET ${PROJECT_NAME})
-
-message("Setup ${TARGET} v${PROJECT_VERSION}")
-
 list(APPEND CMAKE_MODULE_PATH "${CMAKE_CURRENT_LIST_DIR}/cmake")
 
-set(PLUGIN_COMPOSITOR_IMPLEMENTATION "None" CACHE STRING "Defines which implementation is used")
-
-find_package(CompileSettingsDebug CONFIG REQUIRED)
-find_package(${NAMESPACE}Core REQUIRED)
-
-set(CMAKE_POSITION_INDEPENDENT_CODE ON)
-
-if("${PLUGIN_COMPOSITOR_IMPLEMENTATION}" STREQUAL "Nexus")
-    if (COMPOSITORCLIENT_IMPLEMENTATION_PATH)
-        add_subdirectory(${COMPOSITORCLIENT_IMPLEMENTATION_PATH})
-    else()
-        include(GetExternalCode)
-        set(COMPOSITORCLIENT_IMPLEMENTATION_VERSION "master" CACHE STRING "compositorclient implementation version: commit_id_or_tag_or_branch_name")
-        set(COMPOSITORCLIENT_IMPLEMENTATION_REPOSITORY "https://code.rdkcentral.com/r/soc/broadcom/components/rdkcentral/thunder/compositorclient" CACHE STRING "compositorclient implementation repository")
-        GetExternalCode(
-            GIT_REPOSITORY ${COMPOSITORCLIENT_IMPLEMENTATION_REPOSITORY}
-            GIT_VERSION ${COMPOSITORCLIENT_IMPLEMENTATION_VERSION}
-            SOURCE_DIR "Nexus"
-        )
-        add_subdirectory(Nexus)
-    endif()
-else()
-    add_subdirectory (${PLUGIN_COMPOSITOR_IMPLEMENTATION})
-endif()
-
-add_library(${TARGET} SHARED $<TARGET_OBJECTS:${PLUGIN_COMPOSITOR_IMPLEMENTATION}> Module.cpp) 
-
-target_link_libraries(${TARGET}
-    PRIVATE
-        ${PLUGIN_COMPOSITOR_IMPLEMENTATION}
-        ${NAMESPACE}Core::${NAMESPACE}Core
-        CompileSettingsDebug::CompileSettingsDebug
-    )
-
-set(PUBLIC_HEADERS "Client.h")
-
-set_target_properties(${TARGET} PROPERTIES
-    CXX_STANDARD 11
-    CXX_STANDARD_REQUIRED YES
-    FRAMEWORK FALSE
-    PUBLIC_HEADER "${PUBLIC_HEADERS}"
-    LINK_WHAT_YOU_USE TRUE
-    VERSION ${PROJECT_VERSION}
-    SOVERSION ${PROJECT_VERSION_MAJOR}
-)
-
-install(
-    TARGETS ${TARGET} EXPORT ${TARGET}Targets
-    LIBRARY DESTINATION lib COMPONENT libs
-    PUBLIC_HEADER DESTINATION include/${NAMESPACE}/compositor
-    INCLUDES DESTINATION include/${NAMESPACE}
-)
-
-InstallPackageConfig(
-        TARGETS ${TARGET}
-        DESCRIPTION "A ${PLUGIN_COMPOSITOR_IMPLEMENTATION} compositor client implemetation wrapper.")
-
-InstallCMakeConfig(TARGETS ${TARGET})
+add_subdirectory(src)
+add_subdirectory(test)
