+++ Source/compositorclient/include/compositor/Client.h
