+++ Source/compositorclient/src/Module.cpp
