+++ Source/compositorclient/src/Module.h
