+++ Source/compositorclient/src/RPI/CursorData.h
