+++ Source/compositorclient/test/CMakeLists.txt
@@ -1,0 +1,37 @@
+# If not stated otherwise in this file or this component's LICENSE file the
+# following copyright and licenses apply:
+#
+# Copyright 2023 Metrological
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+find_package(CompileSettingsDebug CONFIG REQUIRED)
+find_package(${NAMESPACE}Core CONFIG REQUIRED)
+find_package(${NAMESPACE}Messaging CONFIG REQUIRED)
+find_package(${NAMESPACE}LocalTracer CONFIG REQUIRED)
+
+add_executable(compositorclient_test main.cpp)
+
+target_link_libraries(compositorclient_test
+    PRIVATE
+        ${NAMESPACE}Core::${NAMESPACE}Core
+        ${NAMESPACE}Messaging::${NAMESPACE}Messaging
+        ${NAMESPACE}LocalTracer::${NAMESPACE}LocalTracer
+        CompileSettingsDebug::CompileSettingsDebug
+        compositorclient::compositorclient
+)
+
+if(INSTALL_TESTS)
+    install(TARGETS compositorclient_test DESTINATION bin)
+endif()
